--- backend/src/main/java/com/cookking/services/codeService/CodeService.java
@@ -1,17 +1,0 @@
-package com.cookking.services.codeService;
-
-import com.cookking.models.code.Code;
-import com.cookking.models.code.dto.CreateCodeDto;
-
-import java.util.List;
-
-
-/**
- *  create by kimminsang 2021/09/14
- */
-public interface CodeService {
-    //공통코드 추가
-    public void create(CreateCodeDto createCodeDto);
-    //공통코드 모두 조회
-    public List<Code> findAll();
-}
