+++ backend/src/test/java/com/cookking/controllers/v1/code/CodeControllerTest.java
@@ -35,7 +35,6 @@ public class CodeControllerTest {
     ObjectMapper objectMapper;
 
     private final String URL = "/api/v1/code/";
-    private final Long FIRST_ID = 1L;
 
     private CodeDto getCreateCommonCode(){
         return CodeDto.builder()
